generating reports

+++ src/main/java/org/web/onlineshop/controller/ArticleCategoryController.java
@@ -3,6 +3,8 @@ package org.web.onlineshop.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,9 +16,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.web.onlineshop.dto.ArticleCategoryDto;
+import org.web.onlineshop.exceptions.UnauthorizedAccessException;
 import org.web.onlineshop.model.ArticleCategory;
 import org.web.onlineshop.service.ArticleCategoryService;
 import org.web.onlineshop.util.Constants;
+import org.web.onlineshop.util.UserRole;
 
 @RestController
 @RequestMapping(value = Constants.REST_API_PREFIX + "/article_categories")
@@ -28,9 +32,17 @@ public class ArticleCategoryController
 	@Autowired
 	private ModelMapper modelMapper;
 	
+	@Autowired
+	private HttpServletRequest request;
+	
 	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ArticleCategoryDto> add(@RequestBody ArticleCategoryDto articleCategoryDto)
 	{
+		if (request.getSession().getAttribute("role") != UserRole.ADMINISTRATOR)
+		{
+			throw new UnauthorizedAccessException();
+		}
+		
 		try
 		{
 			ArticleCategory articleCategory = this.modelMapper.map(articleCategoryDto, ArticleCategory.class);
@@ -44,6 +56,11 @@ public class ArticleCategoryController
 	@RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ArticleCategoryDto> update(@RequestBody ArticleCategoryDto articleCategoryDto)
 	{
+		if (request.getSession().getAttribute("role") != UserRole.ADMINISTRATOR)
+		{
+			throw new UnauthorizedAccessException();
+		}
+		
 		try
 		{
 			ArticleCategory articleCategory = this.modelMapper.map(articleCategoryDto, ArticleCategory.class);
@@ -57,6 +74,11 @@ public class ArticleCategoryController
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<?> delete(@PathVariable Long id)
 	{
+		if (request.getSession().getAttribute("role") != UserRole.ADMINISTRATOR)
+		{
+			throw new UnauthorizedAccessException();
+		}
+		
 		try
 		{
 			this.articleCategoryService.delete(id);
