+++ src/main/java/org/web/onlineshop/controller/ArticleController.java
@@ -3,6 +3,8 @@ package org.web.onlineshop.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,12 +17,14 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.web.onlineshop.dto.ArticleDto;
+import org.web.onlineshop.exceptions.UnauthorizedAccessException;
 import org.web.onlineshop.model.Article;
 import org.web.onlineshop.model.ArticleCategory;
 import org.web.onlineshop.service.ArticleCategoryService;
 import org.web.onlineshop.service.ArticleService;
 import org.web.onlineshop.util.ArticleSortType;
 import org.web.onlineshop.util.Constants;
+import org.web.onlineshop.util.UserRole;
 
 @RestController
 @RequestMapping(value = Constants.REST_API_PREFIX + "/articles")
@@ -35,9 +39,17 @@ public class ArticleController
 	@Autowired
 	private ModelMapper modelMapper;
 	
+	@Autowired
+	private HttpServletRequest request;
+	
 	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ArticleDto> add(@RequestBody ArticleDto articleDto)
 	{
+		if (request.getSession().getAttribute("role") != UserRole.ADMINISTRATOR)
+		{
+			throw new UnauthorizedAccessException();
+		}
+		
 		try
 		{
 			Article article = this.modelMapper.map(articleDto, Article.class);
@@ -51,6 +63,11 @@ public class ArticleController
 	@RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ArticleDto> update(@RequestBody ArticleDto articleDto)
 	{
+		if (request.getSession().getAttribute("role") != UserRole.ADMINISTRATOR)
+		{
+			throw new UnauthorizedAccessException();
+		}
+		
 		try
 		{
 			Article article = this.modelMapper.map(articleDto, Article.class);
@@ -64,6 +81,11 @@ public class ArticleController
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<?> delete(@PathVariable Long id)
 	{
+		if (request.getSession().getAttribute("role") != UserRole.ADMINISTRATOR)
+		{
+			throw new UnauthorizedAccessException();
+		}
+		
 		try
 		{
 			this.articleService.delete(id);
